Apply module pattern
